feat: add BookService interface
It is added to provide a layer of abstraction
